Refactor the code mostly renaming
