+++ geode-core/src/distributedTest/java/org/apache/geode/logging/internal/LoggingWithReconnectDistributedTest.java
@@ -105,7 +105,9 @@ public class LoggingWithReconnectDistributedTest implements Serializable {
     locatorDir = temporaryFolder.newFolder(locatorName);
     server1Dir = temporaryFolder.newFolder(server1Name);
     server2Dir = temporaryFolder.newFolder(server2Name);
+
     Invoke.invokeInEveryVM(() -> System.setProperty("jdk.serialFilter", "*"));
+
     int locatorPort = locatorVM.invoke(() -> createLocator());
 
     server1VM.invoke(() -> createServer(server1Name, server1Dir, locatorPort));
