+++ geode-core/src/distributedTest/java/org/apache/geode/management/JmxLocatorReconnectDistributedTest.java
@@ -118,7 +118,9 @@ public class JmxLocatorReconnectDistributedTest implements Serializable {
 
     File locatorDir = temporaryFolder.newFolder(LOCATOR_NAME);
     File serverDir = temporaryFolder.newFolder(SERVER_NAME);
+
     Invoke.invokeInEveryVM(() -> System.setProperty("jdk.serialFilter", "*"));
+
     for (VM vm : asList(locatorVM, serverVM)) {
       vm.invoke(() -> System.setProperty(GEMFIRE_PREFIX + "standard-output-always-on", "true"));
     }
