+++ geode-core/src/integrationTest/java/org/apache/geode/internal/serialization/filter/Java8LocatorLauncherSerializableObjectFilterIntegrationTest.java
@@ -12,23 +12,23 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.io;
+package org.apache.geode.internal.serialization.filter;
 
 import static org.apache.commons.lang3.JavaVersion.JAVA_1_8;
-import static org.apache.commons.lang3.JavaVersion.JAVA_9;
-import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtLeast;
 import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtMost;
 import static org.apache.geode.distributed.ConfigurationProperties.HTTP_SERVICE_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_START;
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_FILE;
-import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPort;
+import static org.apache.geode.distributed.ConfigurationProperties.SERIALIZABLE_OBJECT_FILTER;
+import static org.apache.geode.distributed.ConfigurationProperties.VALIDATE_SERIALIZABLE_OBJECTS;
+import static org.apache.geode.distributed.internal.DistributionConfig.DEFAULT_SERIALIZABLE_OBJECT_FILTER;
+import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPorts;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
 
 import java.io.File;
-import java.io.IOException;
 
 import org.junit.Before;
 import org.junit.Rule;
@@ -36,88 +36,44 @@ import org.junit.Test;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
-import org.apache.geode.distributed.ServerLauncher;
+import org.apache.geode.distributed.LocatorLauncher;
+import org.apache.geode.distributed.internal.DistributionConfig;
+import org.apache.geode.distributed.internal.InternalLocator;
+import org.apache.geode.internal.cache.InternalCache;
+import org.apache.geode.management.ManagementService;
+import org.apache.geode.management.internal.SystemManagementService;
 import org.apache.geode.test.junit.rules.CloseableReference;
 
-public class ServerSetSystemPropertyGlobalSerialFilterIntegrationTest {
+public class Java8LocatorLauncherSerializableObjectFilterIntegrationTest {
 
-  private static final String NAME = "server";
-  private static final String JDK_SERIAL_FILTER_PROPERTY = "jdk.serialFilter";
+  private static final String NAME = "locator";
 
   private File workingDirectory;
+  private int locatorPort;
   private int jmxPort;
 
   @Rule
-  public CloseableReference<ServerLauncher> server = new CloseableReference<>();
+  public CloseableReference<LocatorLauncher> locator = new CloseableReference<>();
   @Rule
   public RestoreSystemProperties restoreSystemProperties = new RestoreSystemProperties();
   @Rule
   public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
   @Before
-  public void setUp() throws IOException {
+  public void setUp() throws Exception {
     workingDirectory = temporaryFolder.newFolder(NAME);
-    jmxPort = getRandomAvailableTCPPort();
+    int[] ports = getRandomAvailableTCPPorts(2);
+    locatorPort = ports[0];
+    jmxPort = ports[1];
   }
 
   @Test
-  public void doesNotChangesEmptyJdkSerialFilter_onJava9orGreater() {
-    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
-
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
-
-    server.set(new ServerLauncher.Builder()
-        .setMemberName(NAME)
-        .setDisableDefaultServer(true)
-        .setWorkingDirectory(workingDirectory.getAbsolutePath())
-        .set(HTTP_SERVICE_PORT, "0")
-        .set(JMX_MANAGER, "true")
-        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
-        .set(JMX_MANAGER_START, "true")
-        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .build())
-        .get()
-        .start();
-
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEmpty();
-  }
-
-  @Test
-  public void doesNotChangeNonEmptySerialFilter_onJava9orGreater() {
-    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
-
-    String existingSerialFilter = "!*";
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, existingSerialFilter);
-
-    server.set(new ServerLauncher.Builder()
-        .setMemberName(NAME)
-        .setDisableDefaultServer(true)
-        .setWorkingDirectory(workingDirectory.getAbsolutePath())
-        .set(HTTP_SERVICE_PORT, "0")
-        .set(JMX_MANAGER, "true")
-        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
-        .set(JMX_MANAGER_START, "true")
-        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .build())
-        .get()
-        .start();
-
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(existingSerialFilter);
-  }
-
-  @Test
-  public void doesNotChangeEmptyJdkSerialFilter_onJava8() {
+  public void configuresValidateSerializableObjects_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
-
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
@@ -128,33 +84,109 @@ public class ServerSetSystemPropertyGlobalSerialFilterIntegrationTest {
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(isValidateSerializableObjectsConfigured())
+        .as(VALIDATE_SERIALIZABLE_OBJECTS)
+        .isTrue();
+  }
+
+  @Test
+  public void usesDefaultSerializableObjectFilter_onJava8() {
+    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
+
+    locator.set(new LocatorLauncher.Builder()
+        .setMemberName(NAME)
+        .setPort(locatorPort)
+        .setWorkingDirectory(workingDirectory.getAbsolutePath())
+        .set(HTTP_SERVICE_PORT, "0")
+        .set(JMX_MANAGER, "true")
+        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
+        .set(JMX_MANAGER_START, "true")
+        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .build())
+        .get()
+        .start();
+
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(getDistributionConfig().getSerializableObjectFilter())
+        .as(SERIALIZABLE_OBJECT_FILTER)
+        .isEqualTo(DEFAULT_SERIALIZABLE_OBJECT_FILTER);
+  }
+
+  @Test
+  public void doesNotChangeEmptySerializableObjectFilter_onJava8() {
+    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
+
+    String emptySerializableObjectFilter = "";
+    locator.set(new LocatorLauncher.Builder()
+        .setMemberName(NAME)
+        .setPort(locatorPort)
+        .setWorkingDirectory(workingDirectory.getAbsolutePath())
+        .set(HTTP_SERVICE_PORT, "0")
+        .set(JMX_MANAGER, "true")
+        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
+        .set(JMX_MANAGER_START, "true")
+        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .set(SERIALIZABLE_OBJECT_FILTER, emptySerializableObjectFilter)
+        .set(VALIDATE_SERIALIZABLE_OBJECTS, "true")
+        .build())
+        .get()
+        .start();
+
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(getDistributionConfig().getSerializableObjectFilter())
+        .as(SERIALIZABLE_OBJECT_FILTER)
         .isEqualTo("");
   }
 
   @Test
-  public void doesNotChangeNonEmptyJdkSerialFilter_onJava8() {
+  public void doesNotChangeNonEmptySerializableObjectFilter_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
-    String existingSerialFilter = "!*";
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, existingSerialFilter);
+    String existingSerializableObjectFilter = "!foo.Bar;*";
 
-    server.set(new ServerLauncher.Builder()
+    locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
-        .setDisableDefaultServer(true)
+        .setPort(locatorPort)
         .setWorkingDirectory(workingDirectory.getAbsolutePath())
         .set(HTTP_SERVICE_PORT, "0")
         .set(JMX_MANAGER, "true")
         .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
         .set(JMX_MANAGER_START, "true")
         .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .set(SERIALIZABLE_OBJECT_FILTER, existingSerializableObjectFilter)
+        .set(VALIDATE_SERIALIZABLE_OBJECTS, "true")
         .build())
         .get()
         .start();
 
-    assertThat(System.getProperty(JDK_SERIAL_FILTER_PROPERTY))
-        .as(JDK_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(existingSerialFilter);
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(getDistributionConfig().getSerializableObjectFilter())
+        .as(SERIALIZABLE_OBJECT_FILTER)
+        .isEqualTo(existingSerializableObjectFilter);
+  }
+
+  private DistributionConfig getDistributionConfig() {
+    InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
+    InternalCache cache = locator.getCache();
+    return cache.getInternalDistributedSystem().getConfig();
+  }
+
+  private SystemManagementService getSystemManagementService() {
+    InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
+    InternalCache cache = locator.getCache();
+    return (SystemManagementService) ManagementService.getManagementService(cache);
+  }
+
+  private boolean isValidateSerializableObjectsConfigured() {
+    return getDistributionConfig().getValidateSerializableObjects();
+  }
+
+  private boolean isJmxManagerStarted() {
+    return getSystemManagementService().isManager();
   }
 }
