+++ geode-core/src/integrationTest/java/org/apache/geode/internal/serialization/filter/Java9LocatorLauncherSerializableObjectFilterIntegrationTest.java
@@ -12,17 +12,18 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.io;
+package org.apache.geode.internal.serialization.filter;
 
-import static org.apache.commons.lang3.JavaVersion.JAVA_1_8;
 import static org.apache.commons.lang3.JavaVersion.JAVA_9;
 import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtLeast;
-import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtMost;
 import static org.apache.geode.distributed.ConfigurationProperties.HTTP_SERVICE_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_START;
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_FILE;
+import static org.apache.geode.distributed.ConfigurationProperties.SERIALIZABLE_OBJECT_FILTER;
+import static org.apache.geode.distributed.ConfigurationProperties.VALIDATE_SERIALIZABLE_OBJECTS;
+import static org.apache.geode.distributed.internal.DistributionConfig.DEFAULT_SERIALIZABLE_OBJECT_FILTER;
 import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPorts;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
@@ -36,23 +37,20 @@ import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
 import org.apache.geode.distributed.LocatorLauncher;
+import org.apache.geode.distributed.internal.DistributionConfig;
 import org.apache.geode.distributed.internal.InternalLocator;
 import org.apache.geode.internal.cache.InternalCache;
-import org.apache.geode.internal.serialization.filter.OpenMBeanFilterPattern;
 import org.apache.geode.management.ManagementService;
 import org.apache.geode.management.internal.SystemManagementService;
 import org.apache.geode.test.junit.rules.CloseableReference;
 
-public class LocatorJmxSerialFilterIntegrationTest {
+public class Java9LocatorLauncherSerializableObjectFilterIntegrationTest {
 
   private static final String NAME = "locator";
-  private static final String JMX_SERIAL_FILTER_PROPERTY =
-      "jmx.remote.rmi.server.serial.filter.pattern";
 
   private File workingDirectory;
   private int locatorPort;
   private int jmxPort;
-  private String openMBeanFilterPattern;
 
   @Rule
   public CloseableReference<LocatorLauncher> locator = new CloseableReference<>();
@@ -67,11 +65,10 @@ public class LocatorJmxSerialFilterIntegrationTest {
     int[] ports = getRandomAvailableTCPPorts(2);
     locatorPort = ports[0];
     jmxPort = ports[1];
-    openMBeanFilterPattern = new OpenMBeanFilterPattern().pattern();
   }
 
   @Test
-  public void configuresJmxSerialFilter_onJava9orGreater() {
+  public void doesNotConfigureValidateSerializableObjects_onJava9orGreater() {
     assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
     locator.set(new LocatorLauncher.Builder()
@@ -89,17 +86,15 @@ public class LocatorJmxSerialFilterIntegrationTest {
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
-        .as(JMX_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(openMBeanFilterPattern);
+    assertThat(isValidateSerializableObjectsConfigured())
+        .as(VALIDATE_SERIALIZABLE_OBJECTS)
+        .isFalse();
   }
 
   @Test
-  public void changesEmptyJmxSerialFilter_onJava9orGreater() {
+  public void doesNotConfigureSerializableObjectFilter_onJava9orGreater() {
     assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
-    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, "");
-
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
         .setPort(locatorPort)
@@ -115,18 +110,16 @@ public class LocatorJmxSerialFilterIntegrationTest {
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
-        .as(JMX_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(openMBeanFilterPattern);
+    assertThat(getDistributionConfig().getSerializableObjectFilter())
+        .as(SERIALIZABLE_OBJECT_FILTER)
+        .isEqualTo(DEFAULT_SERIALIZABLE_OBJECT_FILTER);
   }
 
   @Test
-  public void doesNotChangeNonEmptyJmxSerialFilter_onJava9orGreater() {
+  public void doesNotChangeEmptySerializableObjectFilter_onJava9orGreater() {
     assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
-    String existingJmxSerialFilter = "!*";
-    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, existingJmxSerialFilter);
-
+    String emptySerializableObjectFilter = "";
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
         .setPort(locatorPort)
@@ -136,73 +129,24 @@ public class LocatorJmxSerialFilterIntegrationTest {
         .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
         .set(JMX_MANAGER_START, "true")
         .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .set(SERIALIZABLE_OBJECT_FILTER, emptySerializableObjectFilter)
+        .set(VALIDATE_SERIALIZABLE_OBJECTS, "true")
         .build())
         .get()
         .start();
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
-        .as(JMX_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(existingJmxSerialFilter);
-  }
-
-  @Test
-  public void doesNotConfigureJmxSerialFilter_onJava8() {
-    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
-
-    locator.set(new LocatorLauncher.Builder()
-        .setMemberName(NAME)
-        .setPort(locatorPort)
-        .setWorkingDirectory(workingDirectory.getAbsolutePath())
-        .set(HTTP_SERVICE_PORT, "0")
-        .set(JMX_MANAGER, "true")
-        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
-        .set(JMX_MANAGER_START, "true")
-        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .build())
-        .get()
-        .start();
-
-    assertThat(isJmxManagerStarted())
-        .isTrue();
-    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
-        .as(JMX_SERIAL_FILTER_PROPERTY)
-        .isNull();
-  }
-
-  @Test
-  public void doesNotChangeEmptyJmxSerialFilter_onJava8() {
-    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
-
-    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, "");
-
-    locator.set(new LocatorLauncher.Builder()
-        .setMemberName(NAME)
-        .setPort(locatorPort)
-        .setWorkingDirectory(workingDirectory.getAbsolutePath())
-        .set(HTTP_SERVICE_PORT, "0")
-        .set(JMX_MANAGER, "true")
-        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
-        .set(JMX_MANAGER_START, "true")
-        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .build())
-        .get()
-        .start();
-
-    assertThat(isJmxManagerStarted())
-        .isTrue();
-    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
-        .as(JMX_SERIAL_FILTER_PROPERTY)
+    assertThat(getDistributionConfig().getSerializableObjectFilter())
+        .as(SERIALIZABLE_OBJECT_FILTER)
         .isEqualTo("");
   }
 
   @Test
-  public void doesNotChangeNonEmptyJmxSerialFilter_onJava8() {
-    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
+  public void doesNotChangeNonEmptySerializableObjectFilter_onJava9orGreater() {
+    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
 
-    String existingJmxSerialFilter = "!*";
-    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, existingJmxSerialFilter);
+    String existingSerializableObjectFilter = "!foo.Bar;*";
 
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
@@ -213,21 +157,33 @@ public class LocatorJmxSerialFilterIntegrationTest {
         .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
         .set(JMX_MANAGER_START, "true")
         .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .set(SERIALIZABLE_OBJECT_FILTER, existingSerializableObjectFilter)
+        .set(VALIDATE_SERIALIZABLE_OBJECTS, "true")
         .build())
         .get()
         .start();
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
-        .as(JMX_SERIAL_FILTER_PROPERTY)
-        .isEqualTo(existingJmxSerialFilter);
+    assertThat(getDistributionConfig().getSerializableObjectFilter())
+        .as(SERIALIZABLE_OBJECT_FILTER)
+        .isEqualTo(existingSerializableObjectFilter);
+  }
+
+  private DistributionConfig getDistributionConfig() {
+    InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
+    InternalCache cache = locator.getCache();
+    return cache.getInternalDistributedSystem().getConfig();
   }
 
   private SystemManagementService getSystemManagementService() {
     InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
     InternalCache cache = locator.getCache();
     return (SystemManagementService) ManagementService.getManagementService(cache);
+  }
+
+  private boolean isValidateSerializableObjectsConfigured() {
+    return getDistributionConfig().getValidateSerializableObjects();
   }
 
   private boolean isJmxManagerStarted() {
