+++ geode-core/src/integrationTest/java/org/apache/geode/internal/serialization/filter/LocatorLauncherGlobalSerialFilterBlankIntegrationTest.java
@@ -12,20 +12,14 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.io;
+package org.apache.geode.internal.serialization.filter;
 
-import static org.apache.commons.lang3.JavaVersion.JAVA_1_8;
-import static org.apache.commons.lang3.JavaVersion.JAVA_9;
-import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtLeast;
-import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtMost;
 import static org.apache.geode.distributed.ConfigurationProperties.HTTP_SERVICE_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_START;
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_FILE;
 import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPorts;
-import static org.apache.geode.internal.io.SerialFilterAssertions.assertThatSerialFilterIsNotNull;
-import static org.apache.geode.internal.io.SerialFilterAssertions.assertThatSerialFilterIsNull;
-import static org.assertj.core.api.Assumptions.assumeThat;
+import static org.apache.geode.internal.serialization.filter.SerialFilterAssertions.assertThatSerialFilterIsSameAs;
 
 import java.io.File;
 import java.io.IOException;
@@ -34,16 +28,16 @@ import java.lang.reflect.InvocationTargetException;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
 import org.apache.geode.distributed.LocatorLauncher;
 import org.apache.geode.test.junit.rules.CloseableReference;
 
-public class LocatorGlobalSerialFilterPropertyBlankIntegrationTest {
+public class LocatorLauncherGlobalSerialFilterBlankIntegrationTest {
 
   private static final String NAME = "locator";
-  private static final String JDK_SERIAL_FILTER_PROPERTY = "jdk.serialFilter";
+  private static final ObjectInputFilterApi OBJECT_INPUT_FILTER_API =
+      new ReflectionObjectInputFilterApiFactory().createObjectInputFilterApi();
 
   private File workingDirectory;
   private int locatorPort;
@@ -51,8 +45,6 @@ public class LocatorGlobalSerialFilterPropertyBlankIntegrationTest {
 
   @Rule
   public CloseableReference<LocatorLauncher> locator = new CloseableReference<>();
-  @Rule
-  public RestoreSystemProperties restoreSystemProperties = new RestoreSystemProperties();
   @Rule
   public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -65,11 +57,10 @@ public class LocatorGlobalSerialFilterPropertyBlankIntegrationTest {
   }
 
   @Test
-  public void setsSerialFilterWhenJdkSerialFilterPropertyIsSetToBlank_onJava8()
+  public void doesNotConfigureGlobalSerialFilter_whenFilterIsSetToBlank()
       throws InvocationTargetException, IllegalAccessException {
-    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
-
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
+    Object existingJdkSerialFilter = OBJECT_INPUT_FILTER_API.createFilter(" ");
+    OBJECT_INPUT_FILTER_API.setSerialFilter(existingJdkSerialFilter);
 
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
@@ -83,28 +74,6 @@ public class LocatorGlobalSerialFilterPropertyBlankIntegrationTest {
         .get()
         .start();
 
-    assertThatSerialFilterIsNotNull();
-  }
-
-  @Test
-  public void doesNotSetSerialFilterWhenJdkSerialFilterPropertyIsSetToBlank_onJava9orGreater()
-      throws InvocationTargetException, IllegalAccessException {
-    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
-
-    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
-
-    locator.set(new LocatorLauncher.Builder()
-        .setMemberName(NAME)
-        .setPort(locatorPort)
-        .setWorkingDirectory(workingDirectory.getAbsolutePath())
-        .set(HTTP_SERVICE_PORT, "0")
-        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
-        .set(JMX_MANAGER_START, "true")
-        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .build())
-        .get()
-        .start();
-
-    assertThatSerialFilterIsNull();
+    assertThatSerialFilterIsSameAs(existingJdkSerialFilter);
   }
 }
