+++ geode-core/src/integrationTest/java/org/apache/geode/internal/serialization/filter/LocatorLauncherGlobalSerialFilterPropertyEmptyIntegrationTest.java
@@ -12,14 +12,20 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.io;
+package org.apache.geode.internal.serialization.filter;
 
+import static org.apache.commons.lang3.JavaVersion.JAVA_1_8;
+import static org.apache.commons.lang3.JavaVersion.JAVA_9;
+import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtLeast;
+import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtMost;
 import static org.apache.geode.distributed.ConfigurationProperties.HTTP_SERVICE_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_START;
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_FILE;
 import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPorts;
-import static org.apache.geode.internal.io.SerialFilterAssertions.assertThatSerialFilterIsSameAs;
+import static org.apache.geode.internal.serialization.filter.SerialFilterAssertions.assertThatSerialFilterIsNotNull;
+import static org.apache.geode.internal.serialization.filter.SerialFilterAssertions.assertThatSerialFilterIsNull;
+import static org.assertj.core.api.Assumptions.assumeThat;
 
 import java.io.File;
 import java.io.IOException;
@@ -28,18 +34,16 @@ import java.lang.reflect.InvocationTargetException;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
 import org.apache.geode.distributed.LocatorLauncher;
-import org.apache.geode.internal.serialization.filter.ObjectInputFilterApi;
-import org.apache.geode.internal.serialization.filter.ReflectionObjectInputFilterApiFactory;
 import org.apache.geode.test.junit.rules.CloseableReference;
 
-public class LocatorGlobalSerialFilterExistsIntegrationTest {
+public class LocatorLauncherGlobalSerialFilterPropertyEmptyIntegrationTest {
 
   private static final String NAME = "locator";
-  private static final ObjectInputFilterApi OBJECT_INPUT_FILTER_API =
-      new ReflectionObjectInputFilterApiFactory().createObjectInputFilterApi();
+  private static final String JDK_SERIAL_FILTER_PROPERTY = "jdk.serialFilter";
 
   private File workingDirectory;
   private int locatorPort;
@@ -47,6 +51,8 @@ public class LocatorGlobalSerialFilterExistsIntegrationTest {
 
   @Rule
   public CloseableReference<LocatorLauncher> locator = new CloseableReference<>();
+  @Rule
+  public RestoreSystemProperties restoreSystemProperties = new RestoreSystemProperties();
   @Rule
   public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -59,10 +65,11 @@ public class LocatorGlobalSerialFilterExistsIntegrationTest {
   }
 
   @Test
-  public void doesNotSetSerialFilterWhenSerialFilterExists()
+  public void setsSerialFilterWhenJdkSerialFilterPropertyIsSetToEmpty_onJava8()
       throws InvocationTargetException, IllegalAccessException {
-    Object existingJdkSerialFilter = OBJECT_INPUT_FILTER_API.createFilter("!*");
-    OBJECT_INPUT_FILTER_API.setSerialFilter(existingJdkSerialFilter);
+    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
+
+    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
 
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
@@ -76,6 +83,28 @@ public class LocatorGlobalSerialFilterExistsIntegrationTest {
         .get()
         .start();
 
-    assertThatSerialFilterIsSameAs(existingJdkSerialFilter);
+    assertThatSerialFilterIsNotNull();
+  }
+
+  @Test
+  public void doesNotSetSerialFilterWhenJdkSerialFilterPropertyIsSetToEmpty_onJava9orGreater()
+      throws InvocationTargetException, IllegalAccessException {
+    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
+
+    System.setProperty(JDK_SERIAL_FILTER_PROPERTY, "");
+
+    locator.set(new LocatorLauncher.Builder()
+        .setMemberName(NAME)
+        .setPort(locatorPort)
+        .setWorkingDirectory(workingDirectory.getAbsolutePath())
+        .set(HTTP_SERVICE_PORT, "0")
+        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
+        .set(JMX_MANAGER_START, "true")
+        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .build())
+        .get()
+        .start();
+
+    assertThatSerialFilterIsNull();
   }
 }
