+++ geode-core/src/integrationTest/java/org/apache/geode/internal/serialization/filter/LocatorLauncherJmxSerialFilterIntegrationTest.java
@@ -12,18 +12,17 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.io;
+package org.apache.geode.internal.serialization.filter;
 
 import static org.apache.commons.lang3.JavaVersion.JAVA_1_8;
+import static org.apache.commons.lang3.JavaVersion.JAVA_9;
+import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtLeast;
 import static org.apache.commons.lang3.SystemUtils.isJavaVersionAtMost;
 import static org.apache.geode.distributed.ConfigurationProperties.HTTP_SERVICE_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_PORT;
 import static org.apache.geode.distributed.ConfigurationProperties.JMX_MANAGER_START;
 import static org.apache.geode.distributed.ConfigurationProperties.LOG_FILE;
-import static org.apache.geode.distributed.ConfigurationProperties.SERIALIZABLE_OBJECT_FILTER;
-import static org.apache.geode.distributed.ConfigurationProperties.VALIDATE_SERIALIZABLE_OBJECTS;
-import static org.apache.geode.distributed.internal.DistributionConfig.DEFAULT_SERIALIZABLE_OBJECT_FILTER;
 import static org.apache.geode.internal.AvailablePortHelper.getRandomAvailableTCPPorts;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assumptions.assumeThat;
@@ -37,20 +36,22 @@ import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
 import org.apache.geode.distributed.LocatorLauncher;
-import org.apache.geode.distributed.internal.DistributionConfig;
 import org.apache.geode.distributed.internal.InternalLocator;
 import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.management.ManagementService;
 import org.apache.geode.management.internal.SystemManagementService;
 import org.apache.geode.test.junit.rules.CloseableReference;
 
-public class Java8LocatorSerializableObjectFilterIntegrationTest {
+public class LocatorLauncherJmxSerialFilterIntegrationTest {
 
   private static final String NAME = "locator";
+  private static final String JMX_SERIAL_FILTER_PROPERTY =
+      "jmx.remote.rmi.server.serial.filter.pattern";
 
   private File workingDirectory;
   private int locatorPort;
   private int jmxPort;
+  private String openMBeanFilterPattern;
 
   @Rule
   public CloseableReference<LocatorLauncher> locator = new CloseableReference<>();
@@ -65,10 +66,88 @@ public class Java8LocatorSerializableObjectFilterIntegrationTest {
     int[] ports = getRandomAvailableTCPPorts(2);
     locatorPort = ports[0];
     jmxPort = ports[1];
+    openMBeanFilterPattern = new OpenMBeanFilterPattern().pattern();
   }
 
   @Test
-  public void configuresValidateSerializableObjects_onJava8() {
+  public void configuresJmxSerialFilter_onJava9orGreater() {
+    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
+
+    locator.set(new LocatorLauncher.Builder()
+        .setMemberName(NAME)
+        .setPort(locatorPort)
+        .setWorkingDirectory(workingDirectory.getAbsolutePath())
+        .set(HTTP_SERVICE_PORT, "0")
+        .set(JMX_MANAGER, "true")
+        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
+        .set(JMX_MANAGER_START, "true")
+        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .build())
+        .get()
+        .start();
+
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(openMBeanFilterPattern);
+  }
+
+  @Test
+  public void changesJmxSerialFilter_whenPropertyIsEmpty_onJava9orGreater() {
+    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
+
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, "");
+
+    locator.set(new LocatorLauncher.Builder()
+        .setMemberName(NAME)
+        .setPort(locatorPort)
+        .setWorkingDirectory(workingDirectory.getAbsolutePath())
+        .set(HTTP_SERVICE_PORT, "0")
+        .set(JMX_MANAGER, "true")
+        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
+        .set(JMX_MANAGER_START, "true")
+        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .build())
+        .get()
+        .start();
+
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(openMBeanFilterPattern);
+  }
+
+  @Test
+  public void doesNotChangeJmxSerialFilter_whenPropertyIsNotEmpty_onJava9orGreater() {
+    assumeThat(isJavaVersionAtLeast(JAVA_9)).isTrue();
+
+    String existingJmxSerialFilter = "!*";
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, existingJmxSerialFilter);
+
+    locator.set(new LocatorLauncher.Builder()
+        .setMemberName(NAME)
+        .setPort(locatorPort)
+        .setWorkingDirectory(workingDirectory.getAbsolutePath())
+        .set(HTTP_SERVICE_PORT, "0")
+        .set(JMX_MANAGER, "true")
+        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
+        .set(JMX_MANAGER_START, "true")
+        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
+        .build())
+        .get()
+        .start();
+
+    assertThat(isJmxManagerStarted())
+        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(existingJmxSerialFilter);
+  }
+
+  @Test
+  public void doesNotConfigureJmxSerialFilter_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
     locator.set(new LocatorLauncher.Builder()
@@ -86,14 +165,16 @@ public class Java8LocatorSerializableObjectFilterIntegrationTest {
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(isValidateSerializableObjectsConfigured())
-        .as(VALIDATE_SERIALIZABLE_OBJECTS)
-        .isTrue();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isNull();
   }
 
   @Test
-  public void usesDefaultSerializableObjectFilter_onJava8() {
+  public void doesNotConfigureJmxSerialFilter_whenPropertyIsEmpty_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
+
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, "");
 
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
@@ -110,42 +191,17 @@ public class Java8LocatorSerializableObjectFilterIntegrationTest {
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(getDistributionConfig().getSerializableObjectFilter())
-        .as(SERIALIZABLE_OBJECT_FILTER)
-        .isEqualTo(DEFAULT_SERIALIZABLE_OBJECT_FILTER);
-  }
-
-  @Test
-  public void doesNotChangeEmptySerializableObjectFilter_onJava8() {
-    assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
-
-    locator.set(new LocatorLauncher.Builder()
-        .setMemberName(NAME)
-        .setPort(locatorPort)
-        .setWorkingDirectory(workingDirectory.getAbsolutePath())
-        .set(HTTP_SERVICE_PORT, "0")
-        .set(JMX_MANAGER, "true")
-        .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
-        .set(JMX_MANAGER_START, "true")
-        .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .set(SERIALIZABLE_OBJECT_FILTER, "")
-        .set(VALIDATE_SERIALIZABLE_OBJECTS, "true")
-        .build())
-        .get()
-        .start();
-
-    assertThat(isJmxManagerStarted())
-        .isTrue();
-    assertThat(getDistributionConfig().getSerializableObjectFilter())
-        .as(SERIALIZABLE_OBJECT_FILTER)
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
         .isEqualTo("");
   }
 
   @Test
-  public void doesNotChangeNonEmptySerializableObjectFilter_onJava8() {
+  public void doesNotConfigureJmxSerialFilter_whenPropertyIsNotEmpty_onJava8() {
     assumeThat(isJavaVersionAtMost(JAVA_1_8)).isTrue();
 
-    String existingSerializableObjectFilter = "!foo.Bar;*";
+    String existingJmxSerialFilter = "!*";
+    System.setProperty(JMX_SERIAL_FILTER_PROPERTY, existingJmxSerialFilter);
 
     locator.set(new LocatorLauncher.Builder()
         .setMemberName(NAME)
@@ -156,33 +212,21 @@ public class Java8LocatorSerializableObjectFilterIntegrationTest {
         .set(JMX_MANAGER_PORT, String.valueOf(jmxPort))
         .set(JMX_MANAGER_START, "true")
         .set(LOG_FILE, new File(workingDirectory, NAME + ".log").getAbsolutePath())
-        .set(SERIALIZABLE_OBJECT_FILTER, existingSerializableObjectFilter)
-        .set(VALIDATE_SERIALIZABLE_OBJECTS, "true")
         .build())
         .get()
         .start();
 
     assertThat(isJmxManagerStarted())
         .isTrue();
-    assertThat(getDistributionConfig().getSerializableObjectFilter())
-        .as(SERIALIZABLE_OBJECT_FILTER)
-        .isEqualTo(existingSerializableObjectFilter);
-  }
-
-  private DistributionConfig getDistributionConfig() {
-    InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
-    InternalCache cache = locator.getCache();
-    return cache.getInternalDistributedSystem().getConfig();
+    assertThat(System.getProperty(JMX_SERIAL_FILTER_PROPERTY))
+        .as(JMX_SERIAL_FILTER_PROPERTY)
+        .isEqualTo(existingJmxSerialFilter);
   }
 
   private SystemManagementService getSystemManagementService() {
     InternalLocator locator = (InternalLocator) this.locator.get().getLocator();
     InternalCache cache = locator.getCache();
     return (SystemManagementService) ManagementService.getManagementService(cache);
-  }
-
-  private boolean isValidateSerializableObjectsConfigured() {
-    return getDistributionConfig().getValidateSerializableObjects();
   }
 
   private boolean isJmxManagerStarted() {
