+++ geode-junit/src/integrationTest/java/org/apache/geode/test/junit/internal/JUnitSanctionedSerializablesServiceIntegrationTest.java
@@ -17,7 +17,6 @@ package org.apache.geode.test.junit.internal;
 import org.junit.experimental.categories.Category;
 
 import org.apache.geode.codeAnalysis.SanctionedSerializablesServiceIntegrationTestBase;
-import org.apache.geode.internal.CoreSanctionedSerializablesService;
 import org.apache.geode.internal.serialization.filter.SanctionedSerializablesService;
 import org.apache.geode.test.junit.categories.SanctionedSerializablesTest;
 import org.apache.geode.test.junit.categories.SerializationTest;
@@ -26,7 +25,7 @@ import org.apache.geode.test.junit.categories.SerializationTest;
 public class JUnitSanctionedSerializablesServiceIntegrationTest
     extends SanctionedSerializablesServiceIntegrationTestBase {
 
-  private final SanctionedSerializablesService service = new CoreSanctionedSerializablesService();
+  private final SanctionedSerializablesService service = new JUnitSanctionedSerializablesService();
 
   @Override
   protected SanctionedSerializablesService getService() {
